full refactor extract to separate components. better error handling and descriptions. fix lint errors.
