+++ .gitignore
@@ -22,6 +22,7 @@ wheels/
 *.egg
 
 # Virtual Environment
+.venv/
 venv/
 ENV/
